backend/requirements.txt
django
djangorestframework
channels
channels-redis
daphne
psycopg2-binary
whitenoise
python-dotenv
